Refactor code to add background color and fix formatting of header
Added background color and changed the formatting of the header.

Add relevant information to the HTML file

Added the meta tag for better responsiveness and added links to CSS and favicon files for better presentation of the page. Also included a header tag with a class and associated CSS file.